+++ Xabe.FFmpeg/Conversion/FilterConfiguration.cs
