+++ Xabe.FFmpeg/Conversion/IFilterConfiguration.cs
