+++ Xabe.FFmpeg/FFmpeg.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using Xabe.FFmpeg.Events;
 using Xabe.FFmpeg.Exceptions;
 
 namespace Xabe.FFmpeg
