+++ Xabe.FFmpeg/Helpers/MediaSpeedHelper.cs
